[Refactor] 의미에 맞게 route 관련 변수명 수정 및 초기 state 변수화
