+++ route/Route.js
@@ -6,10 +6,12 @@ export default class Route {
     window.addEventListener("popstate", () => this.route())
     this.route()
   }
+
   route() {
-    const pathname = this.parsePathname()
-    this.changeContent(pathname)
+    const documentId = this.parsePathname()
+    this.changeContent(documentId) //pathname에 따라 Content 컴포넌트를 렌더링함
   }
+
   parsePathname() {
     const { pathname } = location
     return pathname.split("/").at(-1) || "/"
